refactor h1st.contrib.pmfp
