+++ h1st/contrib/pmfp/cli/oracle/exec.py
@@ -16,8 +16,7 @@ from typing import Optional
 import click
 from pandas import Series
 
-from h1st.contrib.pred_maint.models import (BaseFaultPredictor,
-                                            H1ST_BATCH_OUTPUT_S3_DIR_PATH)
+from h1st.contrib.pmfp.models import BaseFaultPredictor, H1ST_BATCH_OUTPUT_S3_DIR_PATH   # noqa: E501
 
 import h1st.utils.debug
 from h1st.utils.path import add_cwd_to_py_path
