+++ h1st/contrib/pmfp/models/base.py
@@ -21,8 +21,8 @@ from h1st.utils.data_proc import ParquetDataset
 from h1st.utils.log import STDOUT_HANDLER
 from h1st.utils import s3
 
-from h1st.contrib.pred_maint.data_mgmt import (EquipmentParquetDataSet,
-                                               EQUIPMENT_INSTANCE_ID_COL, DATE_COL)   # noqa: E501
+from h1st.contrib.pmfp.data_mgmt import (EquipmentParquetDataSet,
+                                         EQUIPMENT_INSTANCE_ID_COL, DATE_COL)
 
 
 __all__ = (
