+++ h1st/contrib/pmfp/models/oracle/teacher/base.py
@@ -10,7 +10,7 @@ from pandas import DataFrame
 
 from h1st.model.oracle.teacher import Teacher
 
-from h1st.contrib.pred_maint.models.base import BaseFaultPredictor
+from h1st.contrib.pmfp.models.base import BaseFaultPredictor
 
 
 class BaseFaultPredTeacher(BaseFaultPredictor, Teacher):
